refactor: int to size_t

+++ get_next_line.h
@@ -16,7 +16,7 @@
 # include <stdlib.h>
 # include <unistd.h>
 
-int		ft_strlen(char *str);
+size_t		ft_strlen(char *str);
 char	*ft_strjoin(char const *s1, char const *s2);
 char	*ft_strndup(char *str, int n);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
